pyproject.toml
[project]
name = "soundmining-monorepo"
version = "2026.03.0"
description = "Monorepo for music composition and DSP"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel",    # Essential for VS Code Notebooks
    "ruff",          # Fast linter/formatter
    "pyyaml",        # For parsing your project manifests
    "pathlib",       # Modern path handling
    "ipywidgets",
    "ipycanvas",
    "matplotlib",
    "soundmining-library"
]

[project.scripts]
sc-start = "tools.sc.launcher:launch_supercollider"
sc-compile = "tools.sc.compiler:main"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["tools", "lib"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


# This ensures that when you run 'uv sync', all your local libs
# are automatically installed in editable mode.
[dependency-groups]
dev = [
    "pytest"
]

[tool.uv.workspace]
members = ["lib/soundmining-library", "projects/*"]

[tool.uv.sources]
# The "Where": This tells uv to look inside the workspace for this package
soundmining-library = { workspace = true }

[tool.ruff]
line-length = 150
target-version = "py312"
# Automatically fix simple linting errors on save
fix = true 

[tool.ruff.lint]
select = ["E", "F", "I"] # Error, Flake8, Isort (clean imports)

